% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterBias.R
\name{clusterBias}
\alias{clusterBias}
\title{Cluster-specific composition bias}
\usage{
clusterBias(counts, sf, clust, threshold = 0.1)
}
\arguments{
\item{counts}{Numeric matrix, count data for genes (in rows) and cells/samples (in columns).}

\item{sf}{Numeric vector, size factors for all cells.}

\item{clust}{Factor, cluster identity for each cell.}

\item{threshold}{Numeric scalar, threshold on the average count for bias estimation.}
}
\value{
Normalization factors representing the scaling bias between clusters.
}
\description{
Compute the composition bias that remains after normalization, based on comparisons between clusters.
}
\details{
We apply \code{sf} to \code{counts} to obtain normalized counts for all cells,
and then compute the average normalized count for each level of \code{clust}.
If cell-specific composition biases were successfully removed by \code{sf}, there should not be any composition biases between clusters.

We use \code{\link{calcNormFactors}} to estimate the remaining bias between clusters.
This is possible as the cluster averages should have few zeroes \emph{and} be very precise (thus reducing the bias of the trimmed mean when DE is imbalanced).
Ideally, the returned values should be equal to unity, indicating that there are no biases in the cluster-specific averages.

Note that \code{lib.size} is set to a constant (of 1 million) to avoid re-introducing composition biases in \code{\link{calcNormFactors}}.
}
\examples{
counts <- matrix(rpois(100000, lambda=2), ncol=100)
sf <- scater::librarySizeFactors(counts)
clust <- sample(3, ncol(counts), replace=TRUE)
clusterBias(counts, sf, clust)
}
\author{
Aaron Lun
}
