% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBatchParam.R
\name{createBatchParam}
\alias{createBatchParam}
\title{Parallelization parameters}
\usage{
createBatchParam(N)
}
\arguments{
\item{N}{Integer, number of workers.}
}
\value{
A \linkS4class{BatchtoolsParam} object deploying to the specified number of workers.
}
\description{
Creates a custom \linkS4class{BatchtoolsParam} object for common use across scripts.
}
\details{
This function assumes a SLURM cluster with the command \code{Rdevel}.
It is intended for development environments where installation of new versions on \code{R} itself would compromise release projects.
}
\author{
Aaron Lun
}
