% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRawMeans.R
\name{generateRawMeans}
\alias{generateRawMeans}
\title{Generate raw means}
\usage{
generateRawMeans(ngenes = 10000, ncells = 200, mode = c("UMI",
  "read"))
}
\arguments{
\item{ngenes}{Integer, number of genes.}

\item{ncells}{Integer, number of cells.}

\item{mode}{String, type of scRNA-seq data.}
}
\value{
A list containing \code{fitted}, a numeric matrix of means for each observation;
\code{mean}, a numeric vector of the gene-specific means;
\code{dispersion}, a numeric vector of the gene-specific dispersions;
and \code{sf}, a numeric vector of the true cell-specific size factors.
}
\description{
Compute the mean expression of each gene across a population of cells, assuming no DE genes.
}
\details{
All counts are presumed to be sampled from a negative binomial (NB) distribution.
The mean of the distribution for any observation is the product of the gene-specific mean and the cell-specific size factor.
The size factor for each cell is sampled from a log-normal distribution.

If \code{mode="UMI"}, we use low gene-specific means and low dispersions.
This mimics what is observed in most UMI data sets, with parameters taken from eyeballing plots of the original inDrop data set.

If \code{mode="read"}, we use high gene-specific means and large dispersions with a decreasing mean-dispersion trend.
This mimics what is observed in many read-based data sets, with parameters taken from eyeballing plots of the ESpresso data set.
}
\examples{
out <- generateRawMeans()
str(out)
}
\author{
Aaron Lun
}
