% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAllMethods.R
\name{runAllMethods}
\alias{runAllMethods}
\title{Run all methods}
\usage{
runAllMethods(counts, threshold = 0)
}
\arguments{
\item{counts}{Numeric matrix, count data with genes in rows and cells/samples in columns.}

\item{threshold}{Numeric scalar, threshold on the mean below which low-abundance genes are to be removed.}
}
\value{
A named list containing numeric vectors of size factors computed from each method.
}
\description{
Compute size factors with a variety of normalization methods.
}
\details{
We apply a variety of methods:
\itemize{
\item TMM normalization with \code{\link{calcNormFactors}}.
\item TMM normalization using an averaged expression profile as the reference sample.
\item DESeq normalization with \code{\link{estimateSizeFactorsForMatrix}}, where all zeroes are ignored during calculation of the geometric mean.
\item DESeq normalization using the arithmetic mean.
\item DESeq normalization with an added pseudo-count of 1.
\item DESeq normalization with a library size-adjusted pseudo-count.
\item Library size normalization with \code{\link{librarySizeFactors}}.
\item Deconvolution with \code{\link{computeSumFactors}} and without any pre-clustering.
\item Deconvolution with \code{\link{computeSumFactors}} and pre-clustering with \code{\link{quickCluster}}.
\item The nearest-neighbor approach in \code{\link{simpleSumFactors}}.
}
For the TMM and DESeq approaches, filtering on \code{threshold} is performed based on the average abundances computed from \code{\link{calcAverage}}.
For the \pkg{scran}-based methods, filtering is done by passing \code{threshold} to the \code{min.mean} arguments.
Library size normalization is unaffected by \code{threshold}.
}
\examples{
counts <- matrix(rpois(100000, lambda=2), ncol=100)
out <- runAllMethods(counts)
str(out)
}
\author{
Aaron Lun
}
