% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCounts.R
\name{sampleCounts}
\alias{sampleCounts}
\title{Sample counts}
\usage{
sampleCounts(means, dispersions)
}
\arguments{
\item{means}{A matrix of observation-specific means.}

\item{dispersions}{A vector of gene-specific dispersions, or a matrix of observation-specific dispersions.}
}
\value{
A numeric matrix of sampled counts with the same dimensions as \code{means}.
}
\description{
Sample count data from a negative binomial (NB) distribution.
}
\details{
This simulates count data from a NB distribution with parameters estimated from \code{\link{generateRawMeans}}.
}
\examples{
out <- generateRawMeans()
counts <- sampleCounts(out$fitted, out$dispersions)
}
\seealso{
\code{\link{generateRawMeans}} to obtain parameters.
}
\author{
Aaron Lun
}
